+++ spec/rspec/rails/matchers/route_to_spec.rb
@@ -3,51 +3,82 @@ require "spec_helper"
 describe "route_to" do
   include RSpec::Rails::Matchers::RoutingMatchers
 
+  it "delegates to assert_recognizes" do
+    self.should_receive(:assert_recognizes).with({ "these" => "options" }, { :method=> :get, :path=>"path" })
+    {:get => "path"}.should route_to("these" => "options")
+  end
+
+  context "with should" do
+    context "when assert_recognizes passes" do
+      it "passes" do
+        self.stub!(:assert_recognizes)
+        expect do
+          {:get => "path"}.should route_to("these" => "options")
+        end.to_not raise_exception
+      end
+    end
+
+    context "when assert_recognizes fails" do
+      it "fails with message from assert_recognizes" do
+        self.stub!(:assert_recognizes) do
+          raise ActiveSupport::TestCase::Assertion.new("this message")
+        end
+        expect do
+          {:get => "path"}.should route_to("these" => "options")
+        end.to raise_error("this message")
+      end
+    end
+
+    context "when an exception is raised" do
+      it "raises that exception" do
+        self.stub!(:assert_recognizes) do
+          raise "oops"
+        end
+        expect do
+          {:get => "path"}.should route_to("these" => "options")
+        end.to raise_exception("oops")
+      end
+    end
+  end
+
+  context "with should_not" do
+    context "when assert_recognizes passes" do
+      it "fails with custom message" do
+        self.stub!(:assert_recognizes)
+        expect do
+          {:get => "path"}.should_not route_to("these" => "options")
+        end.to raise_error(/expected .* not to route to .*/)
+      end
+    end
+
+    context "when assert_recognizes fails" do
+      it "passes" do
+        self.stub!(:assert_recognizes) do
+          raise ActiveSupport::TestCase::Assertion.new("this message")
+        end
+        expect do
+          {:get => "path"}.should_not route_to("these" => "options")
+        end.to_not raise_error
+      end
+    end
+
+    context "when an exception is raised" do
+      it "raises that exception" do
+        self.stub!(:assert_recognizes) do
+          raise "oops"
+        end
+        expect do
+          {:get => "path"}.should_not route_to("these" => "options")
+        end.to raise_exception("oops")
+      end
+    end
+  end
   it "uses failure message from assert_recognizes" do
     self.stub!(:assert_recognizes).and_raise(
-      Test::Unit::AssertionFailedError.new("this message"))
+      ActiveSupport::TestCase::Assertion.new("this message"))
     expect do
       {"this" => "path"}.should route_to("these" => "options")
     end.to raise_error("this message")
   end
 end
 
-describe "be_routable" do
-  include RSpec::Rails::Matchers::RoutingMatchers
-  attr_reader :routes
-
-  before { @routes = double("routes") }
-
-  context "with should" do
-    it "passes if routes recognize the path" do
-      routes.stub(:recognize_path) { {} }
-      expect do
-        {:get => "/a/path"}.should be_routable
-      end.to_not raise_error
-    end
-
-    it "fails if routes do not recognize the path" do
-      routes.stub(:recognize_path) { raise ActionController::RoutingError.new('ignore') }
-      expect do
-        {:get => "/a/path"}.should be_routable
-      end.to raise_error(/expected \{:get=>"\/a\/path"\} to be routable/)
-    end
-  end
-
-  context "with should_not" do
-
-    it "passes if routes do not recognize the path" do
-      routes.stub(:recognize_path) { raise ActionController::RoutingError.new('ignore') }
-      expect do
-        {:get => "/a/path"}.should_not be_routable
-      end.to_not raise_error
-    end
-
-    it "fails if routes recognize the path" do
-      routes.stub(:recognize_path) { {:controller => "foo"} }
-      expect do
-        {:get => "/a/path"}.should_not be_routable
-      end.to raise_error(/expected \{:get=>"\/a\/path"\} not to be routable, but it routes to \{:controller=>"foo"\}/)
-    end
-  end
-end