+++ History.md
@@ -6,6 +6,9 @@
 
 * Enhancements
   * include ApplicationHelper in helper object in helper specs
+  * include request spec extensions in files in spec/integration
+  * include controller spec extensions in groups that use :type => :controller
+    * same for :model, :view, :helper, :mailer, :request, :routing
 
 * Bug fixes
   * restore global config.render_views so you only need to say it once
@@ -13,6 +16,9 @@
   * matchers that delegate to Rails' assertions capture
     ActiveSupport::TestCase::Assertion (so they work properly now with
     should_not in Ruby 1.8.7 and 1.9.1)
+
+* Deprecations
+  * include_self_when_dir_matches
 
 ### 2.3.1 / 2010-12-16